refactor(cli): extract dashboard.py command module from cli monolith
Fifth slice of the cli monolith decomposition. Moves the dashboard
command, its _aggregate_category_totals helper, and all eight
_DASHBOARD_* constants from phi_scan/cli/__init__.py into
phi_scan/cli/dashboard.py. _aggregate_category_totals is re-exported
from phi_scan.cli via __all__ for backwards-compatible test imports.

The dashboard test's four monkeypatch targets (DEFAULT_DATABASE_PATH,
query_recent_scans, get_last_scan, time.sleep) now target
phi_scan.cli.dashboard.* since those names are now read from the
dashboard module. No behavior change.

cli/__init__.py shrinks 894 → 821 lines.
